Rename WIP script to original name
